Refactoring util
