+++ cli/pkg/kctrl/cmd/package/builder/fetch/imgpkg/upstream/upstream.go
@@ -2,7 +2,7 @@ package upstream
 
 import (
 	"fmt"
-	"github.com/vmware-tanzu/carvel-kapp-controller/cli/pkg/kctrl/util"
+	"github.com/vmware-tanzu/carvel-kapp-controller/cli/pkg/kctrl/cmd/package/builder/util"
 	"os"
 	"path/filepath"
 	"strings"
